requirements.txt
numpy
pandas
pyarrow
yfinance
hmmlearn
statsmodels
arch
scikit-learn
fredapi
click
python-dotenv
matplotlib
weasyprint
